+++ src/genome.h
@@ -1,6 +1,6 @@
 #ifndef GENOME_INCLUDE
 #define GENOME_INCLUDE
 
-std::map<std::string, double> create_genome(Rcpp::DataFrame genome) ;
+std::map<std::string, int> create_genome(Rcpp::DataFrame genome) ;
   
 #endif /* GENOME_INCLUDE */
